여러가지수정

+++ src/main/java/com/project/zoopiter/domain/member/dao/MemberDAOImpl.java
@@ -210,8 +210,7 @@ public class MemberDAOImpl implements MemberDAO {
   @Override
   public void delete(String userId) {
     StringBuffer sql = new StringBuffer();
-    sql.append("delete from member ");
-    sql.append(" where user_id = :userId ");
+    sql.append("delete from member where user_id = :userId ");
 
     Map<String, String> param = Map.of("userId", userId);
     template.update(sql.toString(), param);
